Moved the black box example to the stories
